+++ src/main/java/updatetool/common/externalapis/TvdbApi.java
@@ -1,0 +1,3 @@
+package updatetool.common.externalapis;
+
+public interface TvdbApi {}