+++ src/main/java/updatetool/exceptions/ApiCallFailedException.java
@@ -1,6 +1,6 @@
 package updatetool.exceptions;
 
-public class ApiCallFailedException extends Exception {
+public class ApiCallFailedException extends RuntimeException {
     private static final long serialVersionUID = 1L;
 
     public ApiCallFailedException(String msg) {